Update my-css-reset.md
A few spelling and grammar fixes. Rearranged the first paragraph with the two articles and links because English is weird and you have to say 
"... of a couple of"
and because it's nicer for the reader to have access to the articles and their authors in four links in a bullet list than inline.

Awesome article with bad dad jokes and puns!